Added a command-line argument to task files.
Now it is possible to launch the server with a '-a' command line parameter that it is then passed to the task file.

I also debugged some issue in the server related to a variable named 'status' being masked by redefining it.

Using the three text files in the example folder is possible to test the deployment is correct (that -a argument should point to that folder).